+++ CHANGELOG.md
@@ -2,9 +2,10 @@ Version 0.23
 ------------
 - new function `ddlx_createonly()` to create pre-data stuff only. 
 This includes ddl, comments, storage, defaults, settings and owner
-bu not conststraints, indexes, triggers etc. 
+but not conststraints, indexes, triggers etc. 
 Use `ddlx_alter()` to create those after the data is loaded.
 - more use of explicit casts
+- `ddlx_create_trigger()` improvements by PegoraroF10, now with replica and allways triggers
 
 Version 0.22
 ------------